backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
langchain
langchain-openai
openai
python-multipart
numpy
scikit-learn
pandas
yfinance
